menu_tag_project/menu_tag_project/urls.py
"""Модуль корневого маршрутизатора проекта."""
from django.conf import settings
from django.contrib import admin
from django.urls import include, path

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include('displaying_app.urls', namespace='displaying_app')),
]

if settings.DEBUG:
    import debug_toolbar

    urlpatterns += (path('__debug__/', include(debug_toolbar.urls)),)
