menu_tag_project/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-debug-toolbar==5.1.0
sqlparse==0.5.3
tzdata==2025.2
